Create an itk image from vtk image (float type) - build with itk completely without force aot any other problems(it.413)

+++ Header/misMeanFilter.h
@@ -7,12 +7,12 @@ class misMeanFilter
 	{
 	}
 
-  typedef itk::Image< float, 3 > ImageType;
-  typedef   float           InternalPixelType;
+  typedef itk::Image< unsigned short, 3 > ImageType;
+  typedef   unsigned short           InternalPixelType;
 
   typedef itk::Image< InternalPixelType, 3 >  InternalImageType;
   typedef itk::ImageFileReader< InternalImageType > ReaderType;
-  typedef itk::Image< float, 3 > ImageType;
+  typedef itk::Image< unsigned short, 3 > ImageType;
 
   InternalPixelType CalculateMeanValue(int Radius, misMeanFilter::InternalImageType::IndexType index,
   std::string ImageFileName);
